Session level direction attributes should work
Session level direction attributes should work on T.140 data channels where no direction attribute is set on the dcsa(t140) level. 
RFC 4566bis seems to expect that.
It should be discussed what happens if there is a direction attribute on media level in the media declaration for the SCTP declaration..  Probably a similar statement as for session level is needed.